+++ CHANGELOG.md
@@ -3,7 +3,10 @@ VC Verifier Changelog
 
 ## [Unreleased]
 
-### Fixed
+## 3.6.0 (2026-06-18)]
+
+### Added
+
 - JWT presentation holder binding put in place by delegating to `AuthenticationProofPurpose`; linked-data presentations continue to rely on `jsigs.verify()`
 
 ## 3.5.2 (2026-06-02)]